feat: refactor Button component and its usage
