+++ app/shared/components/Footer/FooterContactAndSupport/DonationButton/DonationButton.tsx
@@ -7,28 +7,23 @@ import Button from '~/ds-components/button/Button';
 
 import { type ButtonData } from '~/types/types/common.types';
 
-import { Link } from '~/i18n/navigation';
-import useBreakpoints from '~/shared/hooks/use-breakpoints/useBreakpoints';
+// import { Link } from '~/i18n/navigation';
 
 type DonationDataProps = {
   data: ButtonData;
 };
 
 const DonationButton: React.FC<DonationDataProps> = ({ data }) => {
-  const { isMobile } = useBreakpoints();
-
-  const label = isMobile && data.shortText ? data.shortText : data.text;
-
   return (
-    <Link href={data.link} passHref>
-      <Button
-        size="medium"
-        variant="outlined"
-        color="primary"
-        label={label}
-        startIcon={<SvgImage src="/icons/donation-button.svg" alt="Donation Button" width={24} height={24} />}
-      />
-    </Link>
+    <Button
+      size="medium"
+      variant="outlined"
+      color="primary"
+      label={data.text}
+      shortLabel={data.shortText}
+      link={data.link}
+      startIcon={<SvgImage src="/icons/donation-button.svg" alt="Donation Button" width={24} height={24} />}
+    />
   );
 };
