+++ app/shared/components/Header/SupportButton/SupportButton.tsx
@@ -1,20 +1,15 @@
 import React from 'react';
 
-import type { SupportButtonData } from '~/types/types/header.type';
+import Button from '~/ds-components/button/Button';
 
-import { Link } from '~/i18n/navigation';
-import Button from '~/shared/components/design-system/all-components/button/Button';
+import type { SupportButtonData } from '~/types/types/header.type';
 
 type SupportButtonProps = {
   data: SupportButtonData;
 };
 
 const SupportButton: React.FC<SupportButtonProps> = ({ data }) => {
-  return (
-    <Link href={data.link}>
-      <Button size="medium" variant="contained" color="tertiary" label={data.text}></Button>
-    </Link>
-  );
+  return <Button size="medium" variant="contained" color="tertiary" link={data.link} label={data.text} />;
 };
 
 export default SupportButton;
