+++ app/shared/components/design-system/all-components/button/Button.tsx
@@ -4,6 +4,9 @@ import { Button as MuiButton, ButtonProps as MuiButtonProps, CircularProgress } 
 import { styled } from '@mui/material/styles';
 import { forwardRef, ReactNode } from 'react';
 
+import { Link } from '~/i18n/navigation';
+import useBreakpoints from '~/shared/hooks/use-breakpoints/useBreakpoints';
+
 const CustomButton = styled(MuiButton)({});
 
 type Size = 'large' | 'medium' | 'small';
@@ -15,6 +18,8 @@ type BaseButtonProps = {
   endIcon?: ReactNode;
   loading?: boolean;
   label?: string;
+  shortLabel?: string;
+  link?: string;
 } & (
   | {
       color?: 'primary' | 'secondary';
@@ -28,19 +33,33 @@ type BaseButtonProps = {
 export type ButtonProps = BaseButtonProps & Omit<MuiButtonProps, keyof BaseButtonProps>;
 
 const Button = forwardRef<HTMLButtonElement, ButtonProps>(
-  ({ label, disabled, loading, startIcon, endIcon, children, ...props }, ref) => {
+  ({ label, shortLabel, link, disabled, loading, startIcon, endIcon, children, ...props }, ref) => {
     const isDisabled = disabled ?? loading;
 
+    const { isMobile } = useBreakpoints();
+    label = isMobile && shortLabel ? shortLabel : label;
+
+    const Wrapping = ({ children }: { children: ReactNode }) =>
+      link ? (
+        <Link href={link} passHref>
+          {children}
+        </Link>
+      ) : (
+        <>{children}</>
+      );
+
     return (
-      <CustomButton
-        ref={ref}
-        disabled={isDisabled}
-        startIcon={!loading ? startIcon : undefined}
-        endIcon={!loading ? endIcon : undefined}
-        {...props}
-      >
-        {loading ? <CircularProgress color="inherit" size={25} data-testid="loader" /> : (label ?? children)}
-      </CustomButton>
+      <Wrapping>
+        <CustomButton
+          ref={ref}
+          disabled={isDisabled}
+          startIcon={!loading ? startIcon : undefined}
+          endIcon={!loading ? endIcon : undefined}
+          {...props}
+        >
+          {loading ? <CircularProgress color="inherit" size={25} data-testid="loader" /> : (label ?? children)}
+        </CustomButton>
+      </Wrapping>
     );
   }
 );
