+++ tsconfig.json
@@ -1,11 +1,7 @@
 {
   "compilerOptions": {
     "target": "ES2017",
-    "lib": [
-      "dom",
-      "dom.iterable",
-      "esnext"
-    ],
+    "lib": ["dom", "dom.iterable", "esnext"],
     "allowJs": true,
     "skipLibCheck": true,
     "strict": true,
@@ -27,42 +23,19 @@
     ],
     "baseUrl": ".",
     "paths": {
-      "~/*": [
-        "./app/*"
-      ],
-      "~/public/*": [
-        "./public/*"
-      ],
-      "~/utils/*": [
-        "app/lib/utils/*"
-      ],
-      "~/ds-components/*": [
-        "app/shared/components/design-system/all-components/*"
-      ],
-      "~/components/*": [
-        "app/shared/components/*"
-      ],
-      "~/hooks/*": [
-        "app/shared/hooks/*"
-      ],
-      "~/i18n/*": [
-        "i18n/*"
-      ]
+      "~/*": ["./app/*"],
+      "~/public/*": ["./public/*"],
+      "~/utils/*": ["app/lib/utils/*"],
+      "~/ds-components/*": ["app/shared/components/design-system/all-components/*"],
+      "~/components/*": ["app/shared/components/*"],
+      "~/hooks/*": ["app/shared/hooks/*"],
+      "~/i18n/*": ["i18n/*"]
     },
     "forceConsistentCasingInFileNames": true,
     "noFallthroughCasesInSwitch": true,
     "skipDefaultLibCheck": true,
     "allowSyntheticDefaultImports": true
   },
-  "include": [
-    "**/*.ts",
-    "**/*.tsx",
-    ".next/types/**/*.ts",
-    "jest.config.ts",
-    "next-env.d.ts",
-    "build/types/**/*.ts"
-  ],
-  "exclude": [
-    "node_modules"
-  ]
+  "include": ["**/*.ts", "**/*.tsx", ".next/types/**/*.ts", "jest.config.ts", "next-env.d.ts", "build/types/**/*.ts"],
+  "exclude": ["node_modules"]
 }
